feat: move implementation for k8s to prompt/k8s.go
